+++ apps/load-tests/backend/src/main.ts
@@ -9,7 +9,7 @@ import cluster from "cluster";
 import * as path from "path";
 import { IModelDb, IModelHost, IModelJsFs, SnapshotDb } from "@itwin/core-backend";
 import { Logger, LogLevel } from "@itwin/core-bentley";
-import { BentleyCloudRpcManager, IModelReadRpcInterface, RpcConfiguration, SnapshotIModelRpcInterface } from "@itwin/core-common";
+import { BentleyCloudRpcManager, IModelReadRpcInterface, RpcConfiguration } from "@itwin/core-common";
 import { ECSchemaRpcInterface } from "@itwin/ecschema-rpcinterface-common";
 import { ECSchemaRpcImpl } from "@itwin/ecschema-rpcinterface-impl";
 import { IModelJsExpressServer } from "@itwin/express-server";
@@ -70,7 +70,6 @@ async function initBackend() {
   const rpcConfig = BentleyCloudRpcManager.initializeImpl({ info: { title: "presentation-load-tests-backend", version: "v1.0" } }, [
     ECSchemaRpcInterface,
     IModelReadRpcInterface,
-    SnapshotIModelRpcInterface,
     PresentationRpcInterface,
   ]);
   ECSchemaRpcImpl.register();