+++ packages/components/src/test/propertygrid/UseUnifiedSelection.test.tsx
@@ -47,6 +47,7 @@ describe("usePropertyDataProviderWithUnifiedSelection", () => {
 
     beforeEach(() => {
       selectionHandler = sinon.createStubInstance(SelectionHandler);
+      selectionHandler[Symbol.dispose] = sinon.stub();
     });
 
     it("doesn't set provider keys when handler returns no selection", () => {
@@ -169,6 +170,7 @@ describe("usePropertyDataProviderWithUnifiedSelection", () => {
       });
 
       unmount();
+      // eslint-disable-next-line @typescript-eslint/no-deprecated
       expect(selectionHandler.dispose).to.be.called;
     });
   });