+++ packages/components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log - @itwin/presentation-components
 
+## 5.11.0
+
+### Minor Changes
+
+- [#834](https://github.com/iTwin/presentation/pull/834): Apply and use property value constraints for numeric properties
+- [#834](https://github.com/iTwin/presentation/pull/834): Updated peer dependencies to support iTwin.js Core v5 packages.
+
+### Patch Changes
+
+- [#868](https://github.com/iTwin/presentation/pull/868): Make sure resource that use `dispose` method instead of `Symbol.dispose` are properly disposed.
+- Updated dependencies:
+  - @itwin/presentation-core-interop@1.3.0
+
 ## 5.10.0
 
 ### Minor Changes
@@ -51,6 +64,9 @@
   As the new generation hierarchy building APIs are now available, the old tree-related APIs are now deprecated. See reasoning and migration guide [here](https://github.com/iTwin/presentation/blob/33e79ee8d77f30580a9bab81a72884bda008db25/packages/hierarchies/learning/PresentationRulesMigrationGuide.md).
 
 - [#800](https://github.com/iTwin/presentation/pull/800): Deprecate `viewWithUnifiedSelection` in favor of `enableUnifiedSelectionSyncWithIModel` from `@itwin/unified-selection` package.
+
+  See [iModel selection synchronization with unified selection](https://github.com/iTwin/presentation/blob/master/packages/unified-selection/README.md#imodel-selection-synchronization-with-unified-selection) and [Using with legacy components](https://github.com/iTwin/presentation/blob/master/packages/unified-selection/README.md#imodel-selection-synchronization-with-unified-selection#using-with-legacy-components) for details on how to use `enableUnifiedSelectionSyncWithIModel` in React apps.
+
 - [#802](https://github.com/iTwin/presentation/pull/802): Prefer `Symbol.dispose` over `dispose` for disposable objects.
 
   The package contained a number of types for disposable objects, that had a requirement of `dispose` method being called on them after they are no longer needed. In conjunction with the `using` utility from `@itwin/core-bentley`, usage of such objects looked like this: