+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log - @itwin/presentation-opentelemetry
 
+## 4.2.0
+
+### Minor Changes
+
+- [#834](https://github.com/iTwin/presentation/pull/834): Updated peer dependencies to support iTwin.js Core v5 packages.
+
 ## 4.1.1
 
 ### Patch Changes